Provision always creates keychain when needed
This will default to use the keychain name "build" to create a keychain if importing certs at all for ci provision.  It's always best practice to create a temporary keychain for these cases and use it.  This way, even if no --keychain or --keychain-password are supplied, we use reasonable defaults in CI to create a keychain anyway.